backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
redis
httpx
mutagen
faststream[rabbit]
dotenv
python-multipart
sqlalchemy
alembic
psycopg2-binary
prometheus-fastapi-instrumentator